refactor(quiz): clean up imports and improve code formatting in quiz components
